+++ index.html
@@ -236,5 +236,5 @@
 
 </body>
 <script src="/js/functions.js"></script>
-<script src="/js/main.js"></script>
+<script src="/js/main.js" type="module"></script>
 </html>
