+++ js/data.js
@@ -1,0 +1,17 @@
+const getDataArrayMiniatures = ()=> {
+  const MESSAGES = ['Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+  const USER_NAMES = ['Иван' , 'Петр', 'Милана', 'Глеб', 'Виктория', 'Платон', 'Елена',];
+
+  const PHOTO_DESCRIPTIONS = ['Гостиница', 'Дорога к пляжу', 'Море', 'Девушка', 'Суп', 'Машина', 'Клубника', 'Компот', 'Самолет', 'Обувь', 'Забор', 'Ауди', 'Салат', 'Котик', 'Уги', 'Небо', 'Хор', 'ОлдМобиль', 'Тапочки', 'Пальмы', 'Тайская Еда', 'Закат', 'Краб', 'Концерт', 'Сафари'];
+  return { MESSAGES, USER_NAMES, PHOTO_DESCRIPTIONS };
+};
+
+export { getDataArrayMiniatures };
+
+
